[derive] Rename module ext -> util
This prepares us for a future change which adds more utility
functionality (beyond extension traits) to this module.

gherrit-pr-id: Gcbfd20893d04c5245ea2adf3c1f4ef3ba2327ac6
